finished hw3

+++ src/hw3/Dog.java
@@ -8,13 +8,13 @@ public class Dog extends Animal{
     }
 
     @Override
-    void run(int distance) {
+    void run(double distance) {
         if (distance <= 500) super.run(distance);
         else System.out.println(getName() + " can't run that far.");
     }
 
     @Override
-    void swim(int distance) {
+    void swim(double distance) {
         if (distance <= 10) super.swim(distance);
         else System.out.println(getName() + " can't swim that far.");
     }
